Updates according to PR comments
